Make port 32bit integer

+++ src/inc/librt_session_p.h
@@ -64,14 +64,14 @@ namespace librt
     public:
         SessionPrivate(const std::string &url,
                        const std::string &path,
-                       std::int8_t port,
+                       std::int32_t port,
                        bool authenticationRequired,
                        const std::string &username,
                        const std::string &password);
 
         SessionPrivate(std::string &&url,
                        std::string &&path,
-                       std::int8_t port,
+                       std::int32_t port,
                        bool authenticationRequired,
                        std::string &&username,
                        std::string &&password);
@@ -82,7 +82,7 @@ namespace librt
     private:
         std::string url_;
         std::string path_;
-        std::int8_t port_;
+        std::int32_t port_;
         bool authenticationRequired_;
         std::string username_;
         std::string password_;
